+++ template_test.go
@@ -41,8 +41,8 @@ func TestErrorHandlingEvaluatingTemplate(t *testing.T) {
 				absPath+"unknown-named-slot/index.tw",
 				"parser",
 				fail.ErrSlotNotDefined,
-				"unknown",
 				"user",
+				"unknown",
 			),
 			data: nil,
 		},
@@ -229,6 +229,7 @@ func TestNewTemplate(t *testing.T) {
 		data map[string]any
 		dir  string
 	}{
+		{conf: &config.Config{}, view: "index", data: nil, dir: "slots-optional"},
 		{conf: &config.Config{}, view: "index", data: nil, dir: "reserve-inside-slot"},
 		{conf: &config.Config{}, view: "index", data: nil, dir: "no-stmts"},
 		{conf: &config.Config{}, view: "index", data: nil, dir: "with-inserts"},